Add state-space and output controllers
This supports two types of state-space controllers: StateSpace and
PeriodVariant. The latter allows the user to specify a dt between
controller updates to make the model more accurate.

The output controller classes allow MIMO control without a model.

Examples for a drivetrain, elevator, flywheel, and single-jointed arm
are in the examples folders.